fix: abandoned fractions in favor of decimal
The way fractions are working doesn't allow a lot of sequential operations on probability values to be consistent, because of value limit for numerator and denumerator. Thus, the simplest way to keep precision and stop caring about num. and denum. sizes is to throw them away all together and substitute fractions with simple decimal values. Works for hundreds of dice and seems to be fine for now.

+++ DiceEngine/Utilities/Utils.cs
@@ -5,7 +5,7 @@ public static class Utils {
     /// <param name="a"></param>
     /// <param name="b"></param>
     /// <returns>Greatest common divisor</returns>
-    public static int GCD(int a, int b) {
+    public static long GCD(long a, long b) {
         if (a == 0) {
             return b;
         } else {
